Ask From Multiple Pdfs/requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
huggingface-hub==0.14.1
altair<5

# uncomment to use openai llms
# openai==0.27.6

# uncomment to use instructor embeddings
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2


# Uncomment to use Picone Vectorstore Database
pinecone-client

# uncommit
streamlit_chat
